+++ src/Palette.c
@@ -13,7 +13,6 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "Engine.h"
 #include "Tables.h"
 #include "Tilengine.h"
 
@@ -29,9 +28,9 @@
  */
 TLN_Palette TLN_CreatePalette(int entries) {
   TLN_Palette palette;
-  int size = sizeof(struct Palette) +
-             1024;  // always alloc 256 colors, to avoid crash when tileset uses
-                    // more colors than the palette
+  size_t size = sizeof(struct Palette) +
+                1024;  // always alloc 256 colors, to avoid crash when tileset
+                       // uses more colors than the palette
 
   palette = (TLN_Palette)CreateBaseObject(OT_PALETTE, size);
   if (palette) {
