+++ src/Tilemap.c
@@ -58,8 +58,8 @@ typedef struct {
 TLN_Tilemap TLN_CreateTilemap(int rows, int cols, Tile const *tiles,
                               uint32_t bgcolor, TLN_Tileset tileset) {
   TLN_Tilemap tilemap = NULL;
-  int size =
-      (int)(sizeof(struct Tilemap) + ((long long)rows * cols * sizeof(Tile)));
+  size_t size =
+      sizeof(struct Tilemap) + (size_t)rows * (size_t)cols * sizeof(Tile);
 
   tilemap = (TLN_Tilemap)CreateBaseObject(OT_TILEMAP, size);
   if (!tilemap)
